python-backend/requirements.txt
flask==3.0.0
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
celery==5.3.4
pika==1.3.2
redis==5.0.1
prometheus-client==0.19.0
gunicorn==21.2.0
